// web/src/api/models/quota-update-request.ts
/* tslint:disable */
/* eslint-disable */
/**
 * ATRAG API
 * ATRAG API Documentation
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */



/**
 * Request to update user quotas (supports both single and batch updates)
 * @export
 * @interface QuotaUpdateRequest
 */
export interface QuotaUpdateRequest {
    /**
     * New limit for collection count
     * @type {number}
     * @memberof QuotaUpdateRequest
     */
    'max_collection_count'?: number;
    /**
     * New limit for document count
     * @type {number}
     * @memberof QuotaUpdateRequest
     */
    'max_document_count'?: number;
    /**
     * New limit for documents per collection
     * @type {number}
     * @memberof QuotaUpdateRequest
     */
    'max_document_count_per_collection'?: number;
    /**
     * New limit for bot count
     * @type {number}
     * @memberof QuotaUpdateRequest
     */
    'max_bot_count'?: number;
}