// web/src/api/models/user.ts
/* tslint:disable */
/* eslint-disable */
/**
 * ATRAG API
 * ATRAG API Documentation
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */



/**
 * 
 * @export
 * @interface User
 */
export interface User {
    /**
     * The ID of the user
     * @type {string}
     * @memberof User
     */
    'id'?: string;
    /**
     * The username of the user
     * @type {string}
     * @memberof User
     */
    'username'?: string;
    /**
     * The email of the user
     * @type {string}
     * @memberof User
     */
    'email'?: string;
    /**
     * The role of the user
     * @type {string}
     * @memberof User
     */
    'role'?: string;
    /**
     * Whether the user is active
     * @type {boolean}
     * @memberof User
     */
    'is_active'?: boolean;
    /**
     * The date and time the user joined the system
     * @type {string}
     * @memberof User
     */
    'date_joined'?: string;
    /**
     * The registration source of the user (local, google, github, etc.)
     * @type {string}
     * @memberof User
     */
    'registration_source'?: string;
}