// web/src/api/models/user-quota-info.ts
/* tslint:disable */
/* eslint-disable */
/**
 * ATRAG API
 * ATRAG API Documentation
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


// May contain unused imports in some cases
// @ts-ignore
import type { QuotaInfo } from './quota-info';

/**
 * Complete quota information for a user
 * @export
 * @interface UserQuotaInfo
 */
export interface UserQuotaInfo {
    /**
     * User ID
     * @type {string}
     * @memberof UserQuotaInfo
     */
    'user_id': string;
    /**
     * Username
     * @type {string}
     * @memberof UserQuotaInfo
     */
    'username'?: string;
    /**
     * User email
     * @type {string}
     * @memberof UserQuotaInfo
     */
    'email'?: string;
    /**
     * User role
     * @type {string}
     * @memberof UserQuotaInfo
     */
    'role': string;
    /**
     * List of quota information
     * @type {Array<QuotaInfo>}
     * @memberof UserQuotaInfo
     */
    'quotas': Array<QuotaInfo>;
}